General Trap Fixes
-  Trap Manager has more robust timing to load traps
- Extending Traps is clearer
- Death and Room Duration Changed data types changed to long (as APSlotData exports as longs)

+++ Networking/ArchipelagoSlotData.cs
@@ -12,8 +12,8 @@ namespace Celeste.Mod.CelesteArchipelago
         public long VictoryCondition { get; set; } = 0;
         public long ProgressionSystem { get; set; } = 0;
         public long DisableHeartGates { get; set; } = 0;
-        public int TrapDeathDuration { get; set; } = 10;
-        public int TrapRoomDuration { get; set; } = 3;
+        public long TrapDeathDuration { get; set; } = 10;
+        public long TrapRoomDuration { get; set; } = 3;
 
         private Dictionary<string, PropertyInfo> keyPropertyMap = new Dictionary<string, PropertyInfo>
         {
